properly use settings_tests for tests
