+++ settings_test.py
@@ -1,3 +1,5 @@
+from settings import *  # noqa
+
 import atexit
 import tempfile
 
@@ -27,6 +29,9 @@ def _polite_tmpdir():
     tmp = tempfile.mkdtemp()
     _tmpdirs.add(tmp)
     return tmp
+
+# Make sure the app needed to test translations is present.
+INSTALLED_APPS += TEST_INSTALLED_APPS
 
 # See settings.py for documentation:
 IN_TEST_SUITE = True
@@ -69,30 +74,14 @@ VIDEO_LIBRARIES = ['lib.video.dummy']
 
 ALLOW_SELF_REVIEWS = True
 
-# Make sure debug toolbar output is disabled so it doesn't interfere with any
-# html tests.
-
-
-DEBUG_TOOLBAR_CONFIG = {
-    'INTERCEPT_REDIRECTS': False,
-    'SHOW_TOOLBAR_CALLBACK': lambda r: False,
-    'HIDE_DJANGO_SQL': True,
-    'TAG': 'div',
-    'ENABLE_STACKTRACES': False,
-}
+# Make sure the debug toolbar isn't used during the tests.
+INSTALLED_APPS = [app for app in INSTALLED_APPS if app != 'debug_toolbar']
 
 MOZMARKET_VENDOR_EXCLUDE = []
 
 # These are the default languages. If you want a constrainted set for your
 # tests, you should add those in the tests.
 
-
-def lazy_langs(languages):
-    from product_details import product_details
-    if not product_details.languages:
-        return {}
-    return dict([(i.lower(), product_details.languages[i]['native'])
-                 for i in languages])
 
 AMO_LANGUAGES = (
     'af', 'ar', 'bg', 'ca', 'cs', 'da', 'de', 'el', 'en-US', 'es', 'eu', 'fa',
@@ -116,3 +105,62 @@ ES_DEFAULT_NUM_SHARDS = 3
 
 # Ensure that exceptions aren't re-raised.
 DEBUG_PROPAGATE_EXCEPTIONS = False
+
+
+###############################################################################
+# Only if running on a CI server.
+###############################################################################
+
+if os.environ.get('RUNNING_IN_CI'):
+    import product_details
+
+    LOG_LEVEL = logging.ERROR
+
+    class MockProductDetails:
+        """Main information we need in tests.
+
+        We don't want to rely on the product_details that are automatically
+        downloaded in manage.py for the tests. Also, downloading all the
+        information is very long, and we don't want that for each test build on
+        travis for example.
+
+        So here's a Mock that can be used instead of the real product_details.
+
+        """
+        last_update = False
+        languages = dict((lang, {'native': lang}) for lang in AMO_LANGUAGES)
+        firefox_versions = {"LATEST_FIREFOX_VERSION": "33.1.1"}
+        thunderbird_versions = {"LATEST_THUNDERBIRD_VERSION": "31.2.0"}
+        firefox_history_major_releases = {'1.0': '2004-11-09'}
+
+        def __init__(self):
+            """Some tests need specifics languages.
+
+            This is an excerpt of lib/product_json/languages.json.
+
+            """
+            self.languages.update({
+                u'el': {
+                    u'native':
+                        u'\u0395\u03bb\u03bb\u03b7\u03bd\u03b9\u03ba\u03ac',
+                    u'English': u'Greek'},
+                u'hr': {
+                    u'native': u'Hrvatski',
+                    u'English': u'Croatian'},
+                u'sr': {
+                    u'native': u'\u0421\u0440\u043f\u0441\u043a\u0438',
+                    u'English': u'Serbian'},
+                u'en-US': {
+                    u'native': u'English (US)',
+                    u'English': u'English (US)'},
+                u'tr': {
+                    u'native': u'T\xfcrk\xe7e',
+                    u'English': u'Turkish'},
+                u'cy': {
+                    u'native': u'Cymraeg',
+                    u'English': u'Welsh'},
+                u'sr-Latn': {
+                    u'native': u'Srpski',
+                    u'English': u'Serbian'}})
+
+    product_details.product_details = MockProductDetails()
