+++ tox.ini
@@ -5,7 +5,6 @@ envlist = es, addons-devhub-editors, main, flake8, docs, assets
 basepython = python2.6
 install_command = pip install --no-deps --exists-action=w --download-cache=/tmp/pip-cache --find-links https://pyrepo.addons.mozilla.org/ {opts} {packages}
 setenv =
-    DJANGO_SETTINGS_MODULE=settings_ci
     PYTHONPATH=apps
     DEB_HOST_MULTIARCH=x86_64-linux-gnu  # For M2Crypto on linux.
 whitelist_externals =
